requirements.txt
requests>=2.31.0,<3.0.0
feedparser>=6.0.10,<7.0.0
python-dateutil>=2.8.2,<3.0.0
PyYAML>=6.0.1,<7.0.0
